The fast native implementation (#24)
Developed the native C++ implementation of the whole generation processing of Genetic algorithm.
Also implemented effective C++ format of chromosomes, 2D arrays, etc.
Added native implementation of WorkUnit#_abstract_estimate function as the first native time-resource model.

The whole implementation now scalable by threads (not processes!). It shows the great efficiency by using shared memory.

* Started overall Genetic native implementation

* Introduction to inner C++ chromosome format

* CMake version downgrade

* Converter from inner C++ to Python chromosome format included

* Memory allocation error handling in python_deserializer.cpp

* C++ operators implementation started

* C++ order operators implementation done

* Started resource operators implementation

* Pre-finished genetic C++ scheme

* C++ chromosome evaluator rewrote with inner C++ format

* Ready to run new C++ genetic

* Integrated the first version of native genetic

* Finished first version of Genetic main cycle in native
* Finished sample methods implementation
* Fixed Array2D reallocating issues
* Now full population generating once and propagating to native
- Now recompile.bat not require removal accept

* Debugged and fixed first version of Genetic main cycle in native

* Finished first parallel(!!) version of Genetic main cycle in native

* Chromosome decoding measurement

* Fixed tests collection conflict

+++ sampo/scheduler/genetic/converter.py
@@ -44,7 +44,7 @@ def convert_schedule_to_chromosome(wg: WorkGraph,
 
     # resources for works part of chromosome
     # +1 stores contractors line
-    resource_chromosome = np.zeros((len(order_chromosome), len(worker_name2index) + 1), dtype=int)
+    resource_chromosome = np.zeros((len(order_chromosome), len(worker_name2index) + 1), dtype=np.int32)
 
     for node in order:
         node_id = node.work_unit.id
